Fixes bug where CDAs with inlined sdtc attributes cause crashes
RecordImporter in popHealth needs to register the sdtc namespace
prefix when XML documents are uploaded. health-data-standards
importers expect a Nokogiri document that has all of the namespace
prefixes set up. In the RecordImporter, we were setting up the cda
namespace prefix, but not the sdtc one. I think that on most
documents, we could get away with that because Nokogiri will support
any namespace prefixes that are declared in the start of the document.
By registering the prefix, it fixes this bug.