Allow popHealth to run when not in a git repository
popHealth has code that will check the .git directory for information
about the latest commit and tag. This change will now display
"Unknown" when there is no .git directory. It should also handle the
situation where tags may not be present (this seems to be the case
when testing on Travis CI).