this ties a query to the relationship between a measure and a provider. Previously, a query was a direct child of a measure.
If additional filters are ever used, we'll probably want to make this more universal.

When we can upgrade to Thorax 3, we'll want to use `getQueryForProvider` in the `DashboardSubmeasureView` view as well. Currently, it's too difficult to supply the view with a provider_id, but it'll be easier once https://github.com/walmartlabs/thorax/pull/358 is pulled in