Allows M/R jobs in popHealth to use prefiltering
This pull request allows popHealth to take advantage of the prefiltering support
that has been built into QME and HDS. It is configurable, and on by default, in
config/popHealth.yml.

When prefiltering is enabled, a query is created based on the CQM logic. This
query will be passed to MongoDB as filter for documents to run through the Map
Reduce job. Depending on the measure, this can have a big performance
improvement on measure calculation times.

This commit switches the pointers to HDS and QME in the Gemfile from fixed
versions to their git repos. These should be changed when the gems with the new
functionality are released.