update dashboard measure count
This exposes one of the issues with Thorax... it doesn't have baked-in support for updating small aspects of the UI when properties change. We got around this by using jQuery to update the UI, but this is brittle as it forces us to keep the views up to date when the markup changes, and there aren't any explicit connections between the two files.