wrap long words in the dashboard sidebar
Certain words are too long for the dashboard sidebar, especially NQF numbers that are NamedLikeThis instead of just a number. This will insert soft spaces before/after the slash between the CMS and NQF IDs, as well as allow words to wrap to the next line if they still don't fit.